move ton config to a separate package (fix cycle import)
